backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg==3.2.3
